Moved real_of
