+++ test/sdk/real_of.cpp
@@ -6,7 +6,7 @@
   (See accompanying file LICENSE.md or copy at http://boost.org/LICENSE_1_0.txt)
 */
 //==================================================================================================
-#include <boost/simd/sdk/real_of.hpp>
+#include <boost/simd/meta/real_of.hpp>
 #include <boost/simd/logical.hpp>
 #include <simd_test.hpp>
